Clean human in the loop code
