+++ Tunny/Solver/HumanInTheLoop/HumanSliderInput.cs
@@ -1,33 +1,18 @@
-using System.Diagnostics;
 using System.Drawing;
-using System.IO;
 using System.Text;
-using System.Windows.Forms;
 
 using Python.Runtime;
 
-using Tunny.Handler;
-using Tunny.UI;
-
 namespace Tunny.Solver.HumanInTheLoop
 {
-    public class HumanSliderInput
+    public class HumanSliderInput : HumanInTheLoopBase
     {
         private readonly PyModule _importedLibrary;
         private readonly dynamic _artifactBackend;
-        private readonly dynamic _artifactPath;
-        private readonly string _basePath;
 
-        public HumanSliderInput(string path)
+        public HumanSliderInput(string path) : base(path)
         {
-            PyModule importedLibrary = Py.CreateScope();
-            SetStdOutErrDirection(path, importedLibrary);
-            importedLibrary.Import("os");
-            importedLibrary.Import("textwrap");
-            importedLibrary.Import("threading");
-            importedLibrary.Import("time");
-            importedLibrary.Import("optuna");
-            importedLibrary.Import("optuna_dashboard");
+            PyModule importedLibrary = ImportBaseLibrary();
             importedLibrary.Exec("from optuna_dashboard import save_note");
             importedLibrary.Exec("from optuna_dashboard import SliderWidget");
             importedLibrary.Exec("from optuna_dashboard import ObjectiveUserAttrRef");
@@ -36,65 +21,9 @@ namespace Tunny.Solver.HumanInTheLoop
             importedLibrary.Exec("from optuna_dashboard.artifact import upload_artifact");
             importedLibrary.Exec("from optuna_dashboard.artifact.file_system import FileSystemBackend");
 
-            _basePath = path;
-            _artifactPath = Path.Combine(_basePath, "artifacts");
             dynamic fileSystemBackend = importedLibrary.Get("FileSystemBackend");
             _artifactBackend = fileSystemBackend(base_path: _artifactPath);
             _importedLibrary = importedLibrary;
-        }
-
-        private static void SetStdOutErrDirection(string path, PyModule importedLibrary)
-        {
-            string ioPath = Path.Combine(path, "tmp.out");
-            importedLibrary.Import("sys");
-            importedLibrary.Exec("path = r'" + ioPath + "'");
-            importedLibrary.Exec("sys.stdout = open(path, 'w', encoding='utf-8')");
-            importedLibrary.Exec("sys.stderr = open(path, 'w', encoding='utf-8')");
-        }
-
-        public void WakeOptunaDashboard(Settings.Storage storage)
-        {
-            if (File.Exists(storage.Path) == false)
-            {
-                ResultFileNotExistErrorMessage();
-                return;
-            }
-
-            CheckExistDashboardProcess();
-            string artifactArgument = $"--artifact-dir \"{_artifactPath}\"";
-            var dashboard = new Process();
-            dashboard.StartInfo.FileName = PythonInstaller.GetEmbeddedPythonPath() + @"\Scripts\optuna-dashboard.exe";
-            dashboard.StartInfo.Arguments = storage.GetOptunaStorageCommandLinePathByExtension() + " " + artifactArgument;
-            dashboard.StartInfo.UseShellExecute = false;
-            dashboard.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            dashboard.Start();
-
-            var browser = new Process();
-            browser.StartInfo.FileName = @"http://127.0.0.1:8080/dashboard";
-            browser.StartInfo.UseShellExecute = true;
-            browser.Start();
-        }
-
-        private static void CheckExistDashboardProcess()
-        {
-            Process[] dashboardProcess = Process.GetProcessesByName("optuna-dashboard");
-            if (dashboardProcess.Length > 0)
-            {
-                foreach (Process p in dashboardProcess)
-                {
-                    p.Kill();
-                }
-            }
-        }
-
-        private static void ResultFileNotExistErrorMessage()
-        {
-            TunnyMessageBox.Show(
-                "Please set exist result file path.",
-                "Error",
-                MessageBoxButtons.OK,
-                MessageBoxIcon.Error
-            );
         }
 
         public void SetObjective(dynamic study, string[] objectiveNames)
@@ -131,23 +60,6 @@ namespace Tunny.Solver.HumanInTheLoop
             registerObjectiveFromWidgets(study, widgets: widgets);
         }
 
-        public static int GetRunningTrialNumber(dynamic study)
-        {
-            PyModule lenModule = Py.CreateScope();
-            var pyCode = new StringBuilder();
-            pyCode.AppendLine("import optuna");
-            pyCode.AppendLine("from optuna.trial import TrialState");
-            pyCode.AppendLine("");
-            pyCode.AppendLine("def len4cs(study):");
-            pyCode.AppendLine("    running_trials = study.get_trials(deepcopy=False, states=(TrialState.RUNNING, ))");
-            pyCode.AppendLine("    return len(running_trials)");
-            lenModule.Exec(pyCode.ToString());
-
-            dynamic len = lenModule.Get("len4cs");
-
-            return len(study);
-        }
-
         public void SaveNote(dynamic study, dynamic trial, Bitmap[] bitmaps)
         {
             dynamic uploadArtifact = _importedLibrary.Get("upload_artifact");
@@ -169,20 +81,6 @@ namespace Tunny.Solver.HumanInTheLoop
             dynamic note = textWrap.dedent(noteText.ToString());
             dynamic saveNote = _importedLibrary.Get("save_note");
             saveNote(trial, note);
-        }
-
-        private void CheckDirectoryIsExist()
-        {
-            string artifactPath = Path.Combine(_basePath, "artifacts");
-            string tmpPath = Path.Combine(_basePath, "tmp");
-            if (!Directory.Exists(artifactPath))
-            {
-                Directory.CreateDirectory(artifactPath);
-            }
-            if (!Directory.Exists(tmpPath))
-            {
-                Directory.CreateDirectory(tmpPath);
-            }
         }
     }
 }
